requirements.txt
pandas==2.0.3
plotly==5.15.0
streamlit==1.24.0
ydata-profiling==4.6.1
mysql-connector-python ==8.2.0